move minter files into setup_minter folder
