+++ test-suite/src/common_setup/setup_base_minter.rs
@@ -2,7 +2,6 @@
 // use crate::common_setup::msg::MinterCollectionResponse;
 // use crate::common_setup::msg::MinterSetupParams;
 // use anyhow::Error;
-// use base_factory::state::BaseMinterParams;
 // use cosmwasm_std::{coin, coins, Addr, Timestamp};
 // use cw_multi_test::{AppResponse, Executor};
 // use sg2::msg::CreateMinterMsg;
@@ -14,7 +13,7 @@
 //     state::{ParamsExtension, VendingMinterParams},
 // };
 
-// use super::msg::MinterInstantiateParams;
+// use super::msg::{CodeIds, MinterInstantiateParams};
 
 // // pub const LISTING_FEE: u128 = 0;
 // pub const CREATION_FEE: u128 = 5_000_000_000;
@@ -44,20 +43,31 @@
 //     }
 // }
 
-// pub fn mock_create_minter() -> BaseMinterCreateMsg {
-//     BaseMinterCreateMsg {
-//         init_msg: None,
-//         collection_params: mock_collection_params(),
+// pub fn mock_create_minter(
+//     splits_addr: Option<String>,
+//     collection_params: CollectionParams,
+//     start_time: Option<Timestamp>,
+// ) -> VendingMinterCreateMsg {
+//     VendingMinterCreateMsg {
+//         init_msg: mock_init_extension(splits_addr, start_time),
+//         collection_params,
 //     }
 // }
-// pub fn mock_params() -> BaseMinterParams {
-//     BaseMinterParams {
+
+// pub fn mock_params() -> VendingMinterParams {
+//     VendingMinterParams {
 //         code_id: 1,
 //         creation_fee: coin(CREATION_FEE, NATIVE_DENOM),
 //         min_mint_price: coin(MIN_MINT_PRICE, NATIVE_DENOM),
-//         mint_fee_bps: MINT_FEE_BPS,
+//         mint_fee_bps: MINT_FEE_FAIR_BURN,
 //         max_trading_offset_secs: 60 * 60 * 24 * 7,
-//         extension: None,
+//         extension: ParamsExtension {
+//             max_token_limit: MAX_TOKEN_LIMIT,
+//             max_per_address_limit: MAX_PER_ADDRESS_LIMIT,
+//             airdrop_mint_price: coin(AIRDROP_MINT_PRICE, NATIVE_DENOM),
+//             airdrop_mint_fee_bps: AIRDROP_MINT_FEE_FAIR_BURN,
+//             shuffle_fee: coin(SHUFFLE_FEE, NATIVE_DENOM),
+//         },
 //     }
 // }
 
@@ -140,7 +150,7 @@
 //         .instantiate_contract(
 //             factory_code_id,
 //             minter_admin.clone(),
-//             &base_minter::msg::InstantiateMsg { params },
+//             &vending_factory::msg::InstantiateMsg { params },
 //             &[],
 //             "factory",
 //             None,
@@ -157,7 +167,7 @@
 //     build_collection_response(res, factory_addr)
 // }
 
-// pub fn get_code_ids(router: &mut StargazeApp) -> (u64, u64, u64) {
+// pub fn vending_minter_code_ids(router: &mut StargazeApp) -> CodeIds {
 //     let minter_code_id = router.store_code(contract_minter());
 //     println!("minter_code_id: {}", minter_code_id);
 
@@ -166,7 +176,11 @@
 
 //     let sg721_code_id = router.store_code(contract_sg721());
 //     println!("sg721_code_id: {}", sg721_code_id);
-//     (minter_code_id, factory_code_id, sg721_code_id)
+//     CodeIds {
+//         minter_code_id,
+//         factory_code_id,
+//         sg721_code_id,
+//     }
 // }
 
 // pub fn configure_minter(
@@ -174,9 +188,9 @@
 //     minter_admin: Addr,
 //     collection_params_vec: Vec<CollectionParams>,
 //     minter_instantiate_params_vec: Vec<MinterInstantiateParams>,
+//     code_ids: CodeIds,
 // ) -> Vec<MinterCollectionResponse> {
 //     let mut minter_collection_info: Vec<MinterCollectionResponse> = vec![];
-//     let (minter_code_id, factory_code_id, sg721_code_id) = get_code_ids(app);
 //     for (index, collection_param) in collection_params_vec.iter().enumerate() {
 //         let setup_params: MinterSetupParams = MinterSetupParams {
 //             router: app,
@@ -184,9 +198,9 @@
 //             num_tokens: minter_instantiate_params_vec[index].num_tokens,
 //             collection_params: collection_param.to_owned(),
 //             splits_addr: minter_instantiate_params_vec[index].splits_addr.clone(),
-//             minter_code_id,
-//             factory_code_id,
-//             sg721_code_id,
+//             minter_code_id: code_ids.minter_code_id,
+//             factory_code_id: code_ids.factory_code_id,
+//             sg721_code_id: code_ids.sg721_code_id,
 //             start_time: minter_instantiate_params_vec[index].start_time,
 //             init_msg: minter_instantiate_params_vec[index].init_msg.clone(),
 //         };
