+++ test-suite/src/common_setup/setup_minter/common.rs
@@ -1,0 +1,3 @@
+pub mod constants;
+pub mod minter_params;
+pub mod parse_response;
