+++ test-suite/src/common_setup/setup_minter/common/constants.rs
@@ -1,0 +1,12 @@
+// pub const LISTING_FEE: u128 = 0;
+pub const CREATION_FEE: u128 = 5_000_000_000;
+pub const MINT_PRICE: u128 = 100_000_000;
+
+pub const MAX_TOKEN_LIMIT: u32 = 10000;
+
+pub const MIN_MINT_PRICE: u128 = 50_000_000;
+pub const AIRDROP_MINT_PRICE: u128 = 0;
+pub const MINT_FEE_FAIR_BURN: u64 = 1_000; // 10%
+pub const AIRDROP_MINT_FEE_FAIR_BURN: u64 = 10_000; // 100%
+pub const SHUFFLE_FEE: u128 = 500_000_000;
+pub const MAX_PER_ADDRESS_LIMIT: u32 = 50;
