+++ test-suite/src/common_setup/setup_minter/vending_minter.rs
@@ -1,0 +1,2 @@
+pub mod mock_params;
+pub mod setup;
