+++ test-suite/src/sg_eth_airdrop/setup/configure_mock_minter.rs
@@ -1,7 +1,11 @@
 use crate::common_setup::setup_accounts_and_block::setup_accounts;
 use crate::common_setup::setup_collection_whitelist::setup_whitelist_contract;
-use crate::common_setup::setup_minter::mock_create_minter;
-use crate::common_setup::{contract_boxes::contract_factory, setup_minter::mock_params};
+// use crate::common_setup::setup_minter::mock_create_minter;
+
+use crate::common_setup::contract_boxes::contract_factory;
+use crate::common_setup::setup_minter::vending_minter::mock_params::{
+    mock_create_minter, mock_params,
+};
 use crate::sg_eth_airdrop::constants::collection_constants::CREATION_FEE;
 use crate::sg_eth_airdrop::setup::mock_minter_contract::mock_minter;
 use cosmwasm_std::{coins, Addr, Timestamp};
